third_party/blink/web_tests/platform/linux-chrome/external/wpt/workers/same-site-cookies/first-party.none.tentative.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Check SharedWorker sameSiteCookies option none for first-party
  assert_equals: Worker should get/set SameSite=None cookies only expected "ReadOnLoad:None,ReadOnFetch:None,SetOnRedirectLoad:None,SetOnLoad:None,SetOnRedirectFetch:None,SetOnFetch:None" but got "ReadOnLoad:,ReadOnFetch:,SetOnRedirectLoad:,SetOnLoad:,SetOnRedirectFetch:,SetOnFetch:"
Harness: the test ran to completion.
